+++ tests/test_file_data.py
@@ -7,7 +7,6 @@ from utils.column_values.categorical_column_values import (
     PrimaryServiceType,
     Sector,
 )
-from utils.raw_data_adjustments import RecordsToRemoveInLocationsData
 from utils.validation.validation_rule_custom_type import CustomValidationRules
 from utils.validation.validation_rule_names import RuleNames as RuleName
 
@@ -1034,23 +1033,3 @@ class RawDataAdjustments:
         ("20250101", "12345", "other_data"),
         ("20250101", "67890", "other_data"),
     ]
-
-    locations_data_with_multiple_rows_to_remove = [
-        ("loc_1", "other"),
-        (RecordsToRemoveInLocationsData.dental_practice, "other"),
-        (RecordsToRemoveInLocationsData.dental_practice, "something else"),
-        (RecordsToRemoveInLocationsData.temp_registration, "other"),
-        (RecordsToRemoveInLocationsData.temp_registration, "something else"),
-    ]
-
-    locations_data_with_single_rows_to_remove = [
-        ("loc_1", "other"),
-        (RecordsToRemoveInLocationsData.dental_practice, "other"),
-        (RecordsToRemoveInLocationsData.temp_registration, "other"),
-    ]
-
-    locations_data_without_rows_to_remove = [
-        ("loc_1", "other"),
-    ]
-
-    expected_locations_data = locations_data_without_rows_to_remove